spiders: move Statetul and OAI to common module
