+++ hepcrawl/spiders/base_spider.py
@@ -16,7 +16,7 @@ from urlparse import urljoin
 from scrapy import Request
 from scrapy.spiders import XMLFeedSpider
 
-from . import StatefulSpider
+from .common import StatefulSpider
 from ..items import HEPRecord
 from ..loaders import HEPLoader
 from ..utils import (
