+++ hepcrawl/spiders/brown_spider.py
@@ -19,7 +19,7 @@ from urlparse import urljoin
 from scrapy import Request
 from scrapy.spiders import CrawlSpider
 
-from . import StatefulSpider
+from .common import StatefulSpider
 from ..items import HEPRecord
 from ..loaders import HEPLoader
 from ..utils import (
