+++ hepcrawl/spiders/common/__init__.py
@@ -1,10 +1,13 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of hepcrawl.
-# Copyright (C) 2015, 2016, 2017 CERN.
+# Copyright (C) 2015, 2016, 2017, 2018 CERN.
 #
 # hepcrawl is a free software; you can redistribute it and/or modify it
 # under the terms of the Revised BSD License; see LICENSE file for
 # more details.
 
 from __future__ import absolute_import, division, print_function
+
+from .oaipmh_spider import OAIPMHSpider
+from .stateful_spider import StatefulSpider
