+++ hepcrawl/spiders/edp_spider.py
@@ -19,7 +19,7 @@ from tempfile import mkdtemp
 from scrapy import Request
 from scrapy.spiders import XMLFeedSpider
 
-from . import StatefulSpider
+from .common import StatefulSpider
 from ..extractors.jats import Jats
 from ..items import HEPRecord
 from ..loaders import HEPLoader
