+++ hepcrawl/spiders/iop_spider.py
@@ -18,7 +18,7 @@ from tempfile import mkdtemp
 from scrapy import Request
 from scrapy.spiders import XMLFeedSpider
 
-from . import StatefulSpider
+from .common import StatefulSpider
 from ..extractors.nlm import NLM
 from ..items import HEPRecord
 from ..loaders import HEPLoader
