+++ hepcrawl/spiders/mit_spider.py
@@ -21,7 +21,7 @@ import requests
 from scrapy.http import Request
 from scrapy.spiders import XMLFeedSpider
 
-from . import StatefulSpider
+from .common import StatefulSpider
 from ..items import HEPRecord
 from ..loaders import HEPLoader
 from ..utils import (
