+++ hepcrawl/spiders/phil_spider.py
@@ -17,7 +17,7 @@ from urlparse import urljoin
 from scrapy import Request
 from scrapy.spiders import CrawlSpider
 
-from . import StatefulSpider
+from .common import StatefulSpider
 from ..items import HEPRecord
 from ..loaders import HEPLoader
 from ..utils import (
